[REFACTOR(GameResult)] 패키지 변경
