+++ src/test/kotlin/domain/RefereeTest.kt
@@ -2,6 +2,7 @@ package domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import view.GameResult
 
 class RefereeTest {
     @Test
